// types/account.ts
export enum ACCOUNT_FIELD {
  FIRST_NAME = 'firstName',
  LAST_NAME = 'lastName',
  EMAIL = 'email',
  PASSWORD = 'password',
  CONFIRM_PASSWORD = 'confirm_password',
  ADDRESS = 'address1',
  COUNTRY = 'country',
  CARDNUMBER = 'card_number',
  EXPIRATION = 'expiration',
  CVC = 'cvc',
  OTHER = 'other',
}